reimplemented more efficient way of cleaning old nodes in the hash-table

+++ src/node.h
@@ -61,7 +61,7 @@ private:
     unsigned int nbDirectChildNodes;
 
     float initialValue;
-    size_t numberVisits = 0;
+    int numberVisits = 0;
     std::vector<Node*> childNodes;
 
     Node *parentNode;
